Moving classes and abi to the classes crate.
commit-id:1ba272d3
