+++ crates/bin/starknet-sierra-compile/src/main.rs
@@ -1,9 +1,9 @@
 use std::fs;
 
 use anyhow::Context;
-use cairo_lang_starknet::casm_contract_class::CasmContractClass;
-use cairo_lang_starknet::contract_class::{ContractClass, ContractEntryPoints};
 use cairo_lang_starknet_classes::allowed_libfuncs::ListSelector;
+use cairo_lang_starknet_classes::casm_contract_class::CasmContractClass;
+use cairo_lang_starknet_classes::contract_class::{ContractClass, ContractEntryPoints};
 use cairo_lang_utils::bigint::BigUintAsHex;
 use clap::Parser;
 use serde::Deserialize;
