+++ crates/bin/starknet-sierra-extract-code/Cargo.toml
@@ -8,6 +8,6 @@ description = "Compiler executable for printing Starknet Sierra json included co
 
 [dependencies]
 anyhow.workspace = true
+cairo-lang-starknet-classes = { path = "../../cairo-lang-starknet-classes", version = "2.5.3" }
 clap.workspace = true
 serde_json.workspace = true
-cairo-lang-starknet = { path = "../../cairo-lang-starknet", version = "2.5.3" }
