+++ crates/cairo-lang-starknet-classes/src/casm_contract_class.rs
@@ -1,7 +1,3 @@
-#[cfg(test)]
-#[path = "casm_contract_class_test.rs"]
-mod test;
-
 use cairo_felt::Felt252;
 use cairo_lang_casm::assembler::AssembledCairoProgram;
 use cairo_lang_casm::hints::{Hint, PythonicHint};
@@ -25,8 +21,6 @@ use cairo_lang_sierra_to_casm::compiler::CompilationError;
 use cairo_lang_sierra_to_casm::metadata::{
     calc_metadata, MetadataComputationConfig, MetadataError,
 };
-use cairo_lang_starknet_classes::allowed_libfuncs::AllowedLibfuncsError;
-use cairo_lang_starknet_classes::keccak::starknet_keccak;
 use cairo_lang_utils::bigint::{deserialize_big_uint, serialize_big_uint, BigUintAsHex};
 use cairo_lang_utils::ordered_hash_map::OrderedHashMap;
 use cairo_lang_utils::unordered_hash_set::UnorderedHashSet;
@@ -40,9 +34,15 @@ use serde::{Deserialize, Serialize};
 use starknet_crypto::{poseidon_hash_many, FieldElement};
 use thiserror::Error;
 
+use crate::allowed_libfuncs::AllowedLibfuncsError;
 use crate::compiler_version::{current_compiler_version_id, current_sierra_version_id, VersionId};
 use crate::contract_class::{ContractClass, ContractEntryPoint};
 use crate::felt252_serde::{sierra_from_felt252s, Felt252SerdeError};
+use crate::keccak::starknet_keccak;
+
+#[cfg(test)]
+#[path = "casm_contract_class_test.rs"]
+mod test;
 
 /// The expected gas cost of an entrypoint.
 pub const ENTRY_POINT_COST: i32 = 10000;
