+++ crates/cairo-lang-starknet-classes/src/felt252_serde_test.rs
@@ -7,12 +7,12 @@ use cairo_lang_sierra_generator::canonical_id_replacer::CanonicalReplacer;
 use cairo_lang_sierra_generator::replace_ids::SierraIdReplacer;
 use test_case::test_case;
 
-use super::{sierra_from_felt252s, sierra_to_felt252s, SERDE_SUPPORTED_LONG_IDS};
+use super::{Felt252Serde, SERDE_SUPPORTED_LONG_IDS};
 use crate::compiler_version;
-use crate::felt252_serde::Felt252Serde;
-use crate::test_utils::{get_contract_file_name_from_path, get_example_file_path};
+use crate::felt252_serde::{sierra_from_felt252s, sierra_to_felt252s};
+use crate::test_utils::get_example_file_path;
 
-#[test_case("test_contract::test_contract")]
+#[test_case("test_contract__test_contract")]
 #[test_case("new_syntax_test_contract")]
 #[test_case("hello_starknet")]
 #[test_case("with_erc20")]
@@ -20,14 +20,10 @@ use crate::test_utils::{get_contract_file_name_from_path, get_example_file_path}
 #[test_case("ownable_erc20")]
 #[test_case("upgradable_counter")]
 #[test_case("mintable")]
-#[test_case("multi_component::contract_with_4_components")]
-fn test_felt252_serde(example_contract_path: &str) {
-    let example_file_name = get_contract_file_name_from_path(example_contract_path);
+#[test_case("multi_component__contract_with_4_components")]
+fn test_felt252_serde(name: &str) {
     let sierra = ProgramParser::new()
-        .parse(
-            &read_to_string(get_example_file_path(format!("{example_file_name}.sierra").as_str()))
-                .unwrap(),
-        )
+        .parse(&read_to_string(get_example_file_path(format!("{name}.sierra").as_str())).unwrap())
         .unwrap();
     let replacer = CanonicalReplacer::from_program(&sierra);
     let sierra = replacer.apply(&sierra);
