+++ crates/cairo-lang-starknet/Cargo.toml
@@ -9,7 +9,6 @@ description = "Starknet capabilities and utilities on top of Cairo."
 [dependencies]
 anyhow.workspace = true
 cairo-felt.workspace = true
-cairo-lang-casm = { path = "../cairo-lang-casm", version = "2.5.3", features = ["serde"] }
 cairo-lang-compiler = { path = "../cairo-lang-compiler", version = "2.5.3" }
 cairo-lang-defs = { path = "../cairo-lang-defs", version = "2.5.3" }
 cairo-lang-diagnostics = { path = "../cairo-lang-diagnostics", version = "2.5.3" }
@@ -19,22 +18,16 @@ cairo-lang-plugins = { path = "../cairo-lang-plugins", version = "2.5.3" }
 cairo-lang-semantic = { path = "../cairo-lang-semantic", version = "2.5.3" }
 cairo-lang-sierra = { path = "../cairo-lang-sierra", version = "2.5.3" }
 cairo-lang-sierra-generator = { path = "../cairo-lang-sierra-generator", version = "2.5.3" }
-cairo-lang-sierra-to-casm = { path = "../cairo-lang-sierra-to-casm", version = "2.5.3" }
 cairo-lang-starknet-classes = { path = "../cairo-lang-starknet-classes", version = "2.5.3" }
 cairo-lang-syntax = { path = "../cairo-lang-syntax", version = "2.5.3" }
 cairo-lang-utils = { path = "../cairo-lang-utils", version = "2.5.3", features = ["serde"] }
 const_format.workspace = true
-convert_case.workspace = true
 indoc.workspace = true
 itertools = { workspace = true, default-features = true }
-num-bigint = { workspace = true, default-features = true }
-num-integer.workspace = true
-num-traits = { workspace = true, default-features = true }
 once_cell.workspace = true
 serde = { workspace = true, default-features = true }
 serde_json.workspace = true
 smol_str.workspace = true
-starknet-crypto.workspace = true
 thiserror.workspace = true
 indent.workspace = true
