+++ crates/cairo-lang-starknet/src/lib.rs
@@ -10,12 +10,8 @@ use cairo_lang_semantic::plugin::PluginSuite;
 pub mod abi;
 mod aliased;
 mod analyzer;
-pub mod casm_contract_class;
-pub mod compiler_version;
+pub mod compile;
 pub mod contract;
-pub mod contract_class;
-mod felt252_serde;
-mod felt252_vec_compression;
 pub mod inline_macros;
 pub mod plugin;
