+++ crates/cairo-lang-test-plugin/Cargo.toml
@@ -19,6 +19,7 @@ cairo-lang-semantic = { path = "../cairo-lang-semantic", version = "2.5.3" }
 cairo-lang-sierra = { path = "../cairo-lang-sierra", version = "2.5.3" }
 cairo-lang-sierra-generator = { path = "../cairo-lang-sierra-generator", version = "2.5.3" }
 cairo-lang-starknet = { path = "../cairo-lang-starknet", version = "2.5.3" }
+cairo-lang-starknet-classes = { path = "../cairo-lang-starknet-classes", version = "2.5.3" }
 cairo-lang-syntax = { path = "../cairo-lang-syntax", version = "2.5.3" }
 cairo-lang-utils = { path = "../cairo-lang-utils", version = "2.5.3" }
 indoc.workspace = true
